fix(detection): the class probe's mismatch byte comes from the whole alphabet so full-byte-range pairs are forced to fail, builtins are certified with the flags they ship with, the ground-truth test derives its own stray, and the parse-tree walker is split into a rank-A slots module with import-time opcode tables so no version-only branch exists
