+++ tests/test_sync/test_detection/test_builtin_pattern_safety.py
@@ -7,6 +7,7 @@ from guard_core.handlers.suspatterns_handler import (
 )
 from guard_core.sync.detection_engine.compiler import PatternCompiler
 from guard_core.sync.handlers.suspatterns_handler import (
+    _BUILTIN_PATTERN_COMPILE_FLAGS,
     _DEFAULT_MAX_SCAN_LENGTH,
     _KNOWN_QUADRATIC_BUILTIN_PATTERNS_PENDING_B_XQ_FIX,
     _MEASUREMENT_BORDERLINE_BUILTIN_PATTERNS,
@@ -42,7 +43,9 @@ def test_every_builtin_not_in_the_known_quadratic_set_passes_the_safety_validato
             or pat in _MEASUREMENT_BORDERLINE_BUILTIN_PATTERNS
         ):
             continue
-        ok, reason = pc.validate_pattern_safety(pat)
+        ok, reason = pc.validate_pattern_safety(
+            pat, flags=_BUILTIN_PATTERN_COMPILE_FLAGS
+        )
         if not ok:
             bad.append((cat, reason, pat))
     assert not bad, "built-ins that fail the ReDoS validator:\n" + "\n".join(
